feat(tui): scroll multiline regions to the caret and highlight json
